requirements.txt
fastapi=0.111.0
uvicorn=0.29.0
sentence-transformers=2.7.0
faiss-cpu=1.8.0
fuzzywuzzy=0.18.0
python-Levenshtein=0.25.1
numpy=1.26.4
pydantic=2.7.1
requests=2.31.0